Player can be moved over map via accelerometer
Added a Player object which is controlled by GameInput. Player object
also contains the _tiles array so he can change the color where he
currently is (_tiles is also used to check if he can walk on that tile
(out of bounds...)). Next step: Add AiPlayer objects which are
controlled by an algorithm.